frame-find/scripts/requirements.txt
Flask
Flask-CORS
memvid
PyPDF2
python-dotenv
